+++ src/torch_tools/models/_blocks_2d.py
@@ -15,6 +15,7 @@ from torch_tools.models._argument_processing import (
     process_negative_slope_arg,
     process_boolean_arg,
     process_str_arg,
+    process_2d_block_style_arg,
 )
 
 from torch_tools.torch_utils import disable_biases
@@ -261,9 +262,12 @@ class DownBlock(Sequential):
         The style of the pooling layer to use: can be `"avg"` or `"max"`.
     lr_slope : float
         The negative slope to use in the ``LeakyReLU`` arguments.
-    kernel_size : int
+    kernel_size : int, optional
         The size of the square convolutional kernel to use on the ``Conv2d``
         layers. Must be an odd, positive, int.
+    block_style : str, optional
+        Encoding block style. See
+        ``torch_tools.models._blocks_2d._conv_blocks`` for options.
 
     """
 
@@ -274,6 +278,7 @@ class DownBlock(Sequential):
         pool: str,
         lr_slope: float,
         kernel_size: int = 3,
+        block_style: str = "double_conv",
     ):
         """Build `DownBlock`."""
         super().__init__(
@@ -282,9 +287,9 @@ class DownBlock(Sequential):
                 stride=2,
                 padding=0,
             ),
-            DoubleConvBlock(
-                process_num_feats(in_chans),
-                process_num_feats(out_chans),
+            _conv_blocks[process_2d_block_style_arg(block_style)](
+                in_chans=process_num_feats(in_chans),
+                out_chans=process_num_feats(out_chans),
                 lr_slope=process_negative_slope_arg(lr_slope),
                 kernel_size=process_2d_kernel_size(kernel_size),
             ),
@@ -307,7 +312,10 @@ class UpBlock(Sequential):
         ``ConvTranspose2d`` (``False``).
     lr_slope : float
         Negative slope to use in the ``LeakyReLU`` layers.
-    kernel_size : int
+    block_style : str, optional
+        Encoding block style. See
+        ``torch_tools.models._blocks_2d._conv_blocks`` for options.
+    kernel_size : int, optional
         Size of the covolutional kernel. Must be an odd, positive, int.
 
     """
@@ -318,6 +326,7 @@ class UpBlock(Sequential):
         out_chans: int,
         bilinear: bool,
         lr_slope: float,
+        block_style: str = "double_conv",
         kernel_size: int = 3,
     ):
         """Build `UpBlock`."""
@@ -326,7 +335,7 @@ class UpBlock(Sequential):
                 process_num_feats(in_chans),
                 process_boolean_arg(bilinear),
             ),
-            DoubleConvBlock(
+            _conv_blocks[block_style](
                 process_num_feats(in_chans),
                 process_num_feats(out_chans),
                 lr_slope=process_negative_slope_arg(lr_slope),
@@ -385,6 +394,8 @@ class UNetUpBlock(Module):
     kernel_size : int
         The size of the square convolutional kernel to use in the convolutional
         layers, Should be an odd, positive, int.
+    block_style : str
+        Style of the convolutional block: ``"double_conv"`` or ``"conv_res"``.
 
     """
 
@@ -395,6 +406,7 @@ class UNetUpBlock(Module):
         bilinear: bool,
         lr_slope: float,
         kernel_size: int = 3,
+        block_style: str = "double_conv",
     ):
         """Build `UNetUpBlock`."""
         super().__init__()
@@ -402,12 +414,21 @@ class UNetUpBlock(Module):
         self._out_chans = process_num_feats(out_chans)
 
         self.upsample = self._get_upsampler(process_boolean_arg(bilinear))
-        self.double_conv = DoubleConvBlock(
-            self._in_chans,
-            self._out_chans,
-            lr_slope=lr_slope,
-            kernel_size=process_2d_kernel_size(kernel_size),
-        )
+
+        if process_2d_block_style_arg(block_style) == "double_conv":
+            self.conv_block: Module = DoubleConvBlock(
+                self._in_chans,
+                self._out_chans,
+                lr_slope=lr_slope,
+                kernel_size=process_2d_kernel_size(kernel_size),
+            )
+        else:
+            self.conv_block = ConvResBlock(
+                self._in_chans,
+                self._out_chans,
+                lr_slope,
+                kernel_size,
+            )
 
     @staticmethod
     def _process_in_chans(in_chans: int) -> int:
@@ -580,8 +601,11 @@ class UNetUpBlock(Module):
             height_diff - height_diff // 2,  # Bottom padding
         )
 
-        upsampled = pad(upsampled, padding)
+        upsampled = pad(upsampled, padding)  # pylint: disable=not-callable
 
         # Concatenate along the channel dimension (dim=1) (N, C, H, W)
         concatenated = cat([down_features, upsampled], dim=1)
-        return self.double_conv(concatenated)
+        return self.conv_block(concatenated)
+
+
+_conv_blocks = {"double_conv": DoubleConvBlock, "conv_res": ConvResBlock}