+++ tests/models/test_blocks_2d_call.py
@@ -110,19 +110,16 @@ def test_down_block_call_return_shapes():
     pools = ["max", "avg"]
     lr_slopes = [0.0, 0.1]
     kernel_sizes = [1, 3, 5]
+    block_styles = ["double_conv", "conv_res"]
 
     iterator = product(
-        in_channels,
-        out_channels,
-        pools,
-        lr_slopes,
-        kernel_sizes,
+        in_channels, out_channels, pools, lr_slopes, kernel_sizes, block_styles
     )
 
-    for ins, outs, pool, slope, size in iterator:
+    for ins, outs, pool, slope, size, block_style in iterator:
         batch = rand(10, ins, 50, 100)
 
-        block = DownBlock(ins, outs, pool, slope, size)
+        block = DownBlock(ins, outs, pool, slope, size, block_style)
 
         assert block(batch).shape == (10, outs, 25, 50)
 
@@ -134,16 +131,25 @@ def test_up_block_call_return_shapes():
     bilinears = [True, False]
     slopes = [0.0, 0.1]
     kernels = [1, 3, 5]
+    block_styles = ["double_conv", "conv_res"]
 
-    iterator = product(in_channels, out_channels, bilinears, slopes, kernels)
+    iterator = product(
+        in_channels,
+        out_channels,
+        bilinears,
+        slopes,
+        kernels,
+        block_styles,
+    )
 
-    for in_chans, out_chans, bilinear, lr_slope, kernel_size in iterator:
+    for in_chans, out_chans, bilinear, lr_slope, kernel_size, block_style in iterator:
         block = UpBlock(
             in_chans=in_chans,
             out_chans=out_chans,
             bilinear=bilinear,
             lr_slope=lr_slope,
             kernel_size=kernel_size,
+            block_style=block_style,
         )
         assert block(rand(10, in_chans, 16, 32)).shape == (10, out_chans, 32, 64)
 
@@ -155,16 +161,25 @@ def test_unet_upblock_call_return_shapes():
     bilinears = [True, False]
     slopes = [0.0, 0.1]
     kernels = [1, 3, 5]
+    styles = ["conv_res", "double_conv"]
 
-    iterator = product(in_channels, out_channels, bilinears, slopes, kernels)
+    iterator = product(
+        in_channels,
+        out_channels,
+        bilinears,
+        slopes,
+        kernels,
+        styles,
+    )
 
-    for in_chans, out_chans, bilinear, lr_slope, kernel_size in iterator:
+    for in_chans, out_chans, bilinear, lr_slope, kernel_size, block in iterator:
         block = UNetUpBlock(
             in_chans,
             out_chans,
             bilinear,
             lr_slope,
             kernel_size=kernel_size,
+            block_style=block,
         )
 
         to_upsample = rand(10, in_chans, 25, 50)